+++ src/syslog2irc/processor.py
@@ -7,7 +7,11 @@ syslog2irc.processor
 """
 
 from time import sleep
+from typing import Any, Callable, Iterator, Optional, Set, Tuple
 
+from syslogmp import Message as SyslogMessage
+
+from .router import Router
 from .signals import (
     irc_channel_joined,
     message_approved,
@@ -21,7 +25,13 @@ MESSAGE_TEXT_ENCODING = 'utf-8'
 
 
 class Processor:
-    def __init__(self, router, syslog_message_formatter=None):
+    def __init__(
+        self,
+        router: Router,
+        syslog_message_formatter: Optional[
+            Callable[[SyslogMessage], str]
+        ] = None,
+    ) -> None:
         super(Processor, self).__init__()
         self.router = router
 
@@ -30,12 +40,17 @@ class Processor:
         else:
             self.syslog_message_formatter = format_syslog_message
 
-    def connect_to_signals(self):
+    def connect_to_signals(self) -> None:
         irc_channel_joined.connect(self.router.enable_channel)
         syslog_message_received.connect(self.handle_syslog_message)
         message_received.connect(self.handle_message)
 
-    def handle_syslog_message(self, port, source_address=None, message=None):
+    def handle_syslog_message(
+        self,
+        port: int,
+        source_address: Optional[Tuple[str, int]] = None,
+        message: Optional[SyslogMessage] = None,
+    ) -> None:
         """Process an incoming syslog message."""
         channel_names = self.router.get_channel_names_for_port(port)
 
@@ -50,14 +65,19 @@ class Processor:
         )
 
     def handle_message(
-        self, sender, channel_names=None, text=None, source_address=None
-    ):
+        self,
+        sender: Any,
+        *,
+        channel_names: Optional[Set[str]] = None,
+        text: Optional[str] = None,
+        source_address: Optional[Tuple[str, int]] = None,
+    ) -> None:
         """Process an incoming message."""
         for channel_name in channel_names:
             if self.router.is_channel_enabled(channel_name):
                 message_approved.send(channel_name=channel_name, text=text)
 
-    def run(self, seconds_to_sleep=0.5):
+    def run(self, seconds_to_sleep: float = 0.5) -> None:
         """Run the main loop."""
         try:
             while True:
@@ -68,10 +88,10 @@ class Processor:
         log('Shutting down ...')
 
 
-def format_syslog_message(message):
+def format_syslog_message(message: SyslogMessage) -> str:
     """Format a syslog message to be displayed on IRC."""
 
-    def _generate():
+    def _generate() -> Iterator[str]:
         if message.timestamp is not None:
             timestamp_format = '%Y-%m-%d %H:%M:%S'
             formatted_timestamp = message.timestamp.strftime(timestamp_format)