+++ start-syslog2irc.py
@@ -10,13 +10,15 @@ Receive syslog messages via UDP and show them on IRC.
 :License: MIT, see LICENSE for details.
 """
 
+from typing import Dict, List
+
 from syslog2irc.argparser import parse_args
 from syslog2irc.config import assemble_irc_config
 from syslog2irc.irc import IrcChannel
 from syslog2irc.main import start
 
 
-def start_with_args(routes):
+def start_with_args(routes: Dict[int, List[IrcChannel]]) -> None:
     """Start the IRC bot and the syslog listen server.
 
     All arguments (except for routes) are read from the command line.